pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = ""
authors = [
    {name = "brightertiger",email = "ujjwalsrao@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.25.0",
    "google-adk>=1.21.0",
    "beautifulsoup4>=4.13.0",
    "litellm>=1.80.0",
    "json-repair>=0.54.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.32.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
